Use a unified module registry file, extract dispatcher and configurable classes
