--- src/modules/sigma.classes.camera.js
@@ -1,5 +1,0 @@
-import Camera from "../classes/Camera";
-
-export default function configure(sigma) {
-  sigma.register("sigma.classes.camera", Camera(sigma));
-}
